handlenett-frontend: make item deletable
.. for a users 'own' items.

This became a feature because the partitionkey in db is username.
More easy to delete own elements than others when sending just an
id. By sending an id and wanting to delete others we need to do a lookup
first. Could you say it is a bug? You might. But this is a feature :D